add search bar for movies

+++ app/Http/Controllers/CinemaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Cinema;
 use Illuminate\Http\Request;
+use App\Models\Schedule;
 use App\Exports\CinemaExport;
 use Maatwebsite\Excel\Facades\Excel;
 
@@ -36,7 +37,7 @@ class CinemaController extends Controller
         $request->validate([
             'name' => 'required',
             'location' => 'required|min:10',
-        ],[
+        ], [
             'name.required' => 'Nama bioskop wajib diisi',
             'location.required' => 'Lokasi bioskop wajib diisi',
             'location.min' => 'Lokasi bioskop minimal 10 karakter'
@@ -82,7 +83,7 @@ class CinemaController extends Controller
         $request->validate([
             'name' => 'required',
             'location' => 'required|min:10',
-        ],[
+        ], [
             'name.required' => 'Nama bioskop wajib diisi',
             'location.required' => 'Lokasi bioskop wajib diisi',
             'location.min' => 'Lokasi bioskop minimal 10 karakter'
@@ -104,8 +105,33 @@ class CinemaController extends Controller
      */
     public function destroy($id)
     {
+        $schedules = Schedule::where('cinema_id', $id)->count();
+        if ($schedules) {
+            return redirect()->route('admin.cinemas.index')->with('error', 'Data tidak bisa dihapus karena masih memiliki jadwal');
+        }
+
         Cinema::where('id', $id)->delete();
         return redirect()->route('admin.cinemas.index')->with('success', 'Berhasil menghapus data!');
+    }
+
+        public function trash()
+    {
+        $cinemaTrash = Cinema::onlyTrashed()->get();
+        return view('admin.cinema.trash', compact('cinemaTrash'));
+    }
+
+    public function restore($id)
+    {
+        $cinema = Cinema::onlyTrashed()->find($id);
+        $cinema->restore();
+        return redirect()->route('admin.cinemas.index')->with('success', 'Bioskop Berhasil Dipulihkan');
+    }
+
+    public function deletePermanent($id)
+    {
+        $cinema = Cinema::onlyTrashed()->find($id);
+        $cinema->forceDelete();
+        return redirect()->back()->with('success', 'Bioskop Berhasil Dihapus Seutuhnya');
     }
 
     public function export()
